fix destruction score

+++ src/nl/tue/c2IOE0/group5/ai/GameSimulator.java
@@ -106,7 +106,7 @@ public class GameSimulator {
         simulator.run();
     }
 
-    public float getDestructionScore() {
+    public double getDestructionScore() {
         return gridAnalyzer.getDestructionScore();
     }
